clients: Add quick summarize_square_evictions script for listing pcm.x numbers
This steps from 0% to 200% in increments of 20%, running the square_evictions test program with both
the contracted and expanded set to the number each time.  It outputs a "CSV-like" log file, which
currently demonstrates unexpectedly similar results.  Be sure the core and mask arguments match
(e.g. 0 goes with 0x1, 1 goes with 0x2, and n goes with 2^n).  It doesn't do anything with CAT, so
you might want to first use Intel's pqos tool to assign the target core to its own CoS/cache way.

Here are some improvements we might consider (that, sadly, probably don't account for anomolous
numbers we're seeing):
 - Disable the prefetcher. Or, failing that, use a linear congruential random number generator to
   determine the traversal order, thereby fooling it.
 - Try using Linux hugepages in case TLB misses are somehow affecting the LLC measurements.
 - Add something to force the compiler (and any other that might be used) not to optimize out the
   loops (e.g. a conditional print)