clients/square_evictions: Add -j switch for customizing -l's sampling interval
Recording is now done within the inner (per-pass) loop rather than the outer
one. The default is to only record once per pass---as before---but by passing
-j N, it can instead be made to save an entry when operating on every Nth
location in the array. Corollary: *smaller* numbers mean to sample *faster.*

In practice, to record M times as often as the default (modulus some new
overhead), compute the argument according to the following equation:

               number of cache sets
	    ---------------------------
        N = size of each cache line (B)
	    ---------------------------
	                M

For example, on a machine with 64 B cache lines and 16384 cache lines per way,
to sample at four times the default/legacy rate, N = 16384 / 64 / 4 = 64. Thus,
the astute user would call the program with both -l <outfile> and -j 64.