run-extended-benchmark: Plumb through SIGUSR1 count and square_evictions's -a flag
This was done in preparation for creating a version of run-extended-benchmark that would only run on
a portion of the lockserver's allocation, but that effort has been tabled due to the CSV/JSON effort.
DISCLAIMER: This code is weeks old, and I seem to recall it's not very well tested... My uneducated
hypothesis is that it's better to keep it around than to throw it out, though. Maybe it'll be useful?