jaguar/jaguar: Add trailing[] syntax for appending to an array in set mode
This allows us to dump in a CSV without overwriting the existing contents of an
array-type section; thus, scripts are able to incrementally assemble the dataset
for a particular series. Note that because it's only applicable in the "add-
array-to-array" situation, this syntax can *only* appear at the end of the path
string, and then only in combination with the array input type in set mode!