clients/square_evictions.c: Add -u switch for running forever
When this feature is employed, the contracted phase will be repeated forever and
the expanded case never reached. Believe it or not, this can be spun as a feature:
it enables you to request an allocation that is larger than the working set size
in case that's ever useful to anyone.