clients/lockserver-ng.c: Drop :: in favor of : in optstring
Beware the former feature: whether it populates optarg is based on whether the
optional subargument appears in the *same word* as the switch, which makes it
non-obvious how to extract this argument---which is nonetheless accepted as a
correct invocation---even if it forms its own word.
In the network_rtt experiment, this was causing a pesky issue where the
lockserver ignored the passed-in PID and always signalled sudo instead of the
onmycommand binary.  This took a while to trace back, and I'm still bitter.