network_rtt: Record contender throughput numbers at the very end of the experiment
This is a precaution in case this portion of the experiment ever fails.  For
instance, with very low inter-packet delays, the race between lockserver and
contender initialization causes this data collection to fail; however, in such
situations we currently tend not to be concerned about the contenders anyway,
and should instead ensure at all costs that the *rest* of the data is recorded.