network_rtt: Use line-buffered grep to solve incorrect perf recording interval
Because grep was batching all its output, we weren't running perf for even a second
before the script proceeded to the next line.

The sleep---which was added under suspicion of a race condition but which
unintentionally became simply the duration for which perf samples were actually
recorded---is no longer necessary or desired.