network_rtt: Use latency recording introduced with Mica revision f9bf7ab
This speeds up the grep-based perf recording interval synchronization, limiting
the duration without transaction processing that is logged.

Note that this version of Mica switches to writing the average latency to the
specified outfile instead of to standard output; for this reason, we now need to
search rtt_latencies rather than rtt_client for this information.