data_analysis/mica2_tables: Better method for computing tail latencies
We now simply print the 99.9th-percentile sample from the raw list of points
rather than invoking Dave's ancient make-cdf script.  The old approach was
flawed in that the first line matching 0.999.* wasn't guaranteed not to exceed
the desired percentile by a significant margin.  The new method will work even
if we decide we want to add support for the 99th percentile as well.