driver: Augment the command string by inferring quotes that were probably gobbled by the parent shell
This attempts to resolve the issue where meta.command entries didn't properly reflect argument
quotes that semantically necessary, either for grouping or for protection against evaluation.
The quote insertion is done according to a heuristic: if eval'ing an argument and interpreting it
with printf causes it to no longer lexicographically match, it should probably be quoted.