network_rtt_tail: Cleaner revalloc (-c) implementation to reduce code duplication
We now use a common variable and constants---$cache_ways, $DEFAULT_CACHE_WAYS,
and $DEFAULT_CACHE_WAYS_LOOP---to drive both the alloc (-m) and revalloc (-c)
cases.  This generalization is possible thanks to the new $cache_partition_mite
boolean, which indicates whether the mite is to be granted its own dedicated
cache ways in the basealloc and allocation cases.
Furthermore, cache_ways is initialized to $DEFAULT_CACHE_WAYS at the top of the
script, right after processing the arguments.  As well as being uniform with
most of the other initializations, this approach eliminates the need to
conditionally default-initialize it at the start of the main independent
variable iteration loop.