networked/.sh: Fix default implementation of gencontenderargs
It was mistakenly outputting a single Frankenword instead of three separate ones.  Note that this
function is currently used differently from genserverargs and genclientargs, whose output is eval'ed
and passed to a remote shell, respectively.  Because of this, and unlike the other two functions,
this one should output multiple words that do not contain quotes.