cacheperf: Replace all the fuser bullshit with a *correct* wait-until-perf-exits loop
As it turns out, the original problem was that the perf ps loop ignored the first instance of Perf,
since older kernels had one always-active Perf daemon that no longer exists.  As such, this approach
instead searches for perf instances in sudo process groups in order to maintain compatibility with
both scenarios.
I hesitate to say it, but this actually seems to function *and* be backwards-compatible.