cacheperf: Work around retroactive bug where Perf expected perf.data on stdin
The symptom of this problem is a failure to record results into the Jaguar file,
as well as instances of the following message on stderr and in the log file:

	incompatible file format (rerun with -v to learn more)

This now seems to be happening even when running historical revisions from
before the latest cacheperf reboot (i.e. cachenice_networked and
cachenice_networked_bartender before cachenice_networked_k4.7 was even merged)!
It appears running both Perf 4.4 and Perf 4.7, atop both K4.4 and K4.7.  It is
unknown what caused this change in behavior, but you probably need to
cherry-pick this commit atop old revisions if you're trying to rerun an older
version of the experiment.