jagoff: Allow running multiple instances at once on the same experiment datafile
We now compare md5sums---*after* any path rewriting---to determine whether a
given file clashes with one that we need to extract.  We also use fuser to check
whether we should nuke a file when we're done with it, which ensures that the
last instance reading the Perf data corresponding to a particular datapoint
automagically "garbage collects" that Perf data file!