networked: Fix bug where implicit variables weren't actually handled
When adding the implicit variable code to the module system base, I based it on
the implicit function code, which used type to check whether functions were
defined.  Obviously, that doesn't work for variables, so every such variable
went undetected by this logic.  The practical upshot of this was that the gzip
module would report an internal error about EXPECTS_FILES when you tried to run
it.