networked: Add mica_wss_percentage submodule that can be used to fix table_entries by percentage of the allocation
This supports experiments crafted in response to concerns that we were actually
varying two things at once in our cache ways sensitivity analysis experiments:
the number of cache ways and whether the working/hot set could fit in cache.

To use this feature, simply include it in the module list you pass on the
command line, in *addition* to the mica module.