network_rtt: Temporarily stub out perf recording
Upon doing this, we see a noticeable decrease in noise, indicating that perf's inconsistent
performance footprint was likely responsible for many of our bad results.
(Note that we're moving the recording to before the warmup interval instead of merely
eradicating it in order to break the counter summary and perf.data saving logic at the end.